hoist routes for mute timings to facilitate react-router upgrade
