+++ public/app/features/alerting/unified/components/mute-timings/NewMuteTiming.tsx
@@ -1,0 +1,20 @@
+import { withErrorBoundary } from '@grafana/ui';
+
+import { AlertmanagerPageWrapper } from '../AlertingPageWrapper';
+
+import MuteTimingForm from './MuteTimingForm';
+
+const NewMuteTimingPage = () => (
+  <AlertmanagerPageWrapper
+    navId="am-routes"
+    pageNav={{
+      id: 'alert-policy-new',
+      text: 'Add mute timing',
+    }}
+    accessType="notification"
+  >
+    <MuteTimingForm />
+  </AlertmanagerPageWrapper>
+);
+
+export default withErrorBoundary(NewMuteTimingPage, { style: 'page' });
